backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
argon2-cffi==23.1.0
PyJWT==2.8.0
python-dotenv==1.0.1
